Move `CallStack` into `Program`
I am working on changes that require access to the call stack to test,
similar to how tests can already access the operands. Moving the call
stack into `Program` seems like a good step towards this, and increases
the consistency between the call stack and other runtime data.

This is not purely a refactoring. It has the effect of no longer
re-starting a program when calling `run`. Instead, it will continue the
program where it previously left of.

This new behavior is so far untested, but that is going to change. I am
already working on features that are going to rely on this behavior.