Decouple jumping and pushing return address
Convert the former call instructions into equivalent jump instruction
which do not push a return address, and add a dedicated instruction for
that.

This could be seen as a cleanup, since pushing return addresses, which
was previously duplicated between the two call instructions, is now
handled by dedicated code. Though the new instruction requires a bit of
additional complexity, which makes this less of a clear win.

It does make the instruction set much more flexible though, which the
upcoming implementation of tail call optimization is going to require.