Rename `InvalidAddress` to `InvalidOperand`
I'm about to add more cases of invalid operands, and I assume that the
function that triggers the effect already makes it clear what the reason
for the effect is.

If that holds true, than a single effect is more than enough, and
probably preferable to adding new effects for all the different cases.